refactor: restructure homepage components by moving client and main page files to a new directory
